chore: move tab components to Tabs folder
