+++ src/components/NewTab/Tabs/CloseBtn.tsx
