+++ src/components/NewTab/Tabs/StarBtn.tsx
